+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 21 August 2017
+- Breaking changes to the /api/v2/user endpoint. preferences[setup] now returns a string if the user is not setup correctly.
+
 ## 18 August 2017
 - Added email_blacklist attribute to google_integrations.
 - Breaking changes to the /api/v2/contacts/duplicates endpoints. Duplicates are now resources with uuids.